Remove unused stuff from previous TOP implementation
