+++ lib/process/action/process.ex
@@ -1,7 +1,5 @@
 defmodule Helix.Process.Action.Process do
 
-  import HELL.Macros
-
   alias Helix.Entity.Model.Entity
   alias Helix.Entity.Query.Entity, as: EntityQuery
   alias Helix.Network.Model.Connection
@@ -12,9 +10,6 @@ defmodule Helix.Process.Action.Process do
   alias Helix.Process.Internal.Process, as: ProcessInternal
   alias Helix.Process.Model.Process
   alias Helix.Process.Model.Processable
-  alias Helix.Process.Query.Process, as: ProcessQuery
-  alias Helix.Process.State.TOP.Manager, as: ManagerTOP
-  alias Helix.Process.State.TOP.Server, as: ServerTOP
 
   alias Helix.Process.Event.Process.Completed, as: ProcessCompletedEvent
   alias Helix.Process.Event.Process.Created, as: ProcessCreatedEvent
@@ -60,11 +55,12 @@ defmodule Helix.Process.Action.Process do
     ProcessInternal.delete(process)
 
     event =
-      # if reason == :completed do
+      if reason == :completed do
         ProcessCompletedEvent.new(process)
-      # else
+      else
+        ProcessCompletedEvent.new(process)
         # ProcessKilledEvent.new(process)
-      # end
+      end
 
     {:ok, [event]}
   end
@@ -106,108 +102,10 @@ defmodule Helix.Process.Action.Process do
   # defp signal_handler(:SIGFILED, process, %{file: file}),
   #   do: Processable.file_deleted(process.data, process, file)
 
-  @spec pause(Process.t) ::
-    :ok
-  @doc """
-  Changes a process state to _paused_
-
-  This will cause it not to "process" the allocated resources and thus suspend
-  it's effect.
-
-  Some processes might still consume resources (without any progress) on paused
-  state
-
-  This function is idempotent
-
-  ### Examples
-
-      iex> pause(%Process{})
-      :ok
-  """
-  def pause(process) do
-  end
-
-  @spec resume(Process.t) ::
-    :ok
-  @doc """
-  Changes a process state from _paused_ to _running_
-
-  This will allow the process to continue processing resources and causing side-
-  effects
-
-  This function is idempotent
-
-  ### Examples
-
-      iex> resume(%Process{})
-      :ok
-  """
-  def resume(process) do
-  end
-
-  @spec priority(Process.t, 0..5) ::
-    :ok
-  @doc """
-  Changes the priority of a process
-
-  Effectively this will change how much proportionally the input process will
-  receive of dynamic resources. The higher the priority, the higher the amount
-  of dynamic resources a process will receive
-
-  ### Examples
-
-      iex> priority(%Process{}, 1)
-      :ok
-
-      iex> priority(%Process{}, 5)
-      :ok
-  """
-  def priority(process, priority) when priority in 0..5 do
-  end
-
-  @spec kill(Process.t, atom) ::
-    :ok
-  @doc """
-  Stops a process with reason `reason`
-
-  ### Examples
-
-      iex> kill(%Process{}, :normal)
-      :ok
-  """
-  def kill(process, _reason) do
-  end
-
-  @doc false
-  def reset_processes_on_server(gateway_id) do
-    case ManagerTOP.get(gateway_id) do
-      nil ->
-        :noop
-      pid ->
-        processes = ProcessQuery.get_processes_on_server(gateway_id)
-        ServerTOP.reset_processes(pid, processes)
-    end
-  end
-
   @spec prepare_create_params(base_params, Entity.id) ::
     Process.create_params
   defp prepare_create_params(params, source_entity_id),
     do: Map.put(params, :source_entity_id, source_entity_id)
-
-  @spec get_process_entities(base_params) ::
-    {source_entity :: Entity.id, target_entity :: Entity.id}
-  defp get_process_entities(params) do
-    source_entity = EntityQuery.fetch_by_server(params.gateway_id)
-
-    target_entity =
-      if params.gateway_id == params.target_id do
-        source_entity
-      else
-        EntityQuery.fetch_by_server(params.target_id)
-      end
-
-    {source_entity.entity_id, target_entity.entity_id}
-  end
 
   @spec get_process_ips(base_params) ::
     {gateway_ip :: Network.ip, target_ip :: Network.ip}
