+++ lib/process/event/handler/process.ex
@@ -3,7 +3,6 @@ defmodule Helix.Process.Event.Handler.Process do
   alias Helix.Event
   alias Helix.Process.Action.Process, as: ProcessAction
 
-  alias Helix.Process.Event.Process.Completed, as: ProcessCompletedEvent
   alias Helix.Process.Event.Process.Signaled, as: ProcessSignaledEvent
 
   def signal_handler(event = %ProcessSignaledEvent{}) do
